fix: try to fix undo delete logic

+++ features/content/kol/src/main/java/com/tokopedia/kol/feature/comment/view/fragment/KolCommentNewFragment.kt
@@ -9,8 +9,12 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import androidx.fragment.app.FragmentActivity
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.LifecycleOwner
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.google.android.material.snackbar.Snackbar
 import com.tokopedia.abstraction.base.view.adapter.Visitable
 import com.tokopedia.abstraction.base.view.fragment.BaseDaggerFragment
 import com.tokopedia.abstraction.common.utils.image.ImageHandler
@@ -160,26 +164,51 @@ class KolCommentNewFragment : BaseDaggerFragment(), KolComment.View, KolComment.
     ): Boolean {
         if (canDeleteComment || isInfluencer()) {
             if (userSession != null && userSession?.isLoggedIn != false) {
-                deleteComment(adapterPosition)
-                var toBeDeleted = true
-                view?.let {
-                    val coroutineScope = CoroutineScope(Dispatchers.Main)
-                    coroutineScope.launch {
-                        if (activity != null && isAdded) {
-                            if (toBeDeleted)
-                                presenter.deleteComment(id, adapterPosition)
-                        }
-                    }
-                    Toaster.toasterCustomCtaWidth =
-                            com.tokopedia.unifyprinciples.R.dimen.unify_space_96
-                    Toaster.build(
-                            it,
-                            getString(R.string.kol_delete_1_comment),
-                            TOASTER_DURATION_DELETE_COMMENT,
-                            Toaster.TYPE_NORMAL,
-                    ).show()
+                val view = this.view ?: return false
 
+                var toBeDeleted = true
+                deleteComment(adapterPosition)
+
+                Toaster.toasterCustomCtaWidth = com.tokopedia.unifyprinciples.R.dimen.unify_space_96
+                val toaster = Toaster.build(
+                    view,
+                    getString(R.string.kol_delete_1_comment),
+                    TOASTER_DURATION_DELETE_COMMENT,
+                    Toaster.TYPE_NORMAL,
+                    getString(R.string.kol_delete_comment_ok)
+                ) {
+                    feedAnalytics.clickKembalikanCommentPage(postId, authorId, isVideoPost, isFollowed, postType)
+                    adapter?.clearList()
+                    presenter.getCommentFirstTime(arguments?.getInt(ARGS_ID) ?: 0)
+                    toBeDeleted = false
                 }
+
+                val deleteFn = {
+                    if (activity != null && isAdded) {
+                        if (toBeDeleted) presenter.deleteComment(id, adapterPosition)
+                        toBeDeleted = false
+                    }
+                }
+
+                val lifecycleObserver = object : LifecycleEventObserver {
+                    override fun onStateChanged(source: LifecycleOwner, event: Lifecycle.Event) {
+                        if (event != Lifecycle.Event.ON_DESTROY) return
+                        deleteFn()
+                    }
+                }
+                viewLifecycleOwner.lifecycle.addObserver(lifecycleObserver)
+
+                val callback = object : Snackbar.Callback() {
+                    override fun onDismissed(transientBottomBar: Snackbar?, event: Int) {
+                        deleteFn()
+                        viewLifecycleOwner.lifecycle.removeObserver(lifecycleObserver)
+                        transientBottomBar?.removeCallback(this)
+                    }
+                }
+                Toaster.snackBar.addCallback(callback)
+
+                toaster.show()
+
             } else {
                 goToLogin()
             }
